refactor: move embedded utility functions to iframe.utils and update imports
